requirements.txt
netmiko>=4.1.2
textfsm>=1.1.3
nettoolkit>=0.1.0
pyyaml>=6.0.1
sqlalchemy>=2.0.0
pandas>=2.0.0
click>=8.1.7
rich>=13.7.0 